Add some tests for when differential equations are deactivated.
