+++ idaes/core/solvers/features.py
@@ -206,116 +206,6 @@ def dae():
         0.4873531310307455e-2,
     )
 
-def dae_with_non_time_indexed_constraint():
-    """This provides a DAE model for solver testing. This model contains a non-
-    time-indexed variable and constraint.
-
-    The problem and expected result are from A test problem from
-    https://archimede.dm.uniba.it/~testset/report/chemakzo.pdf.
-
-    Args:
-        None
-
-    Returns:
-        (tuple): Pyomo ConcreteModel, correct solved value for y[1] to y[5] and y6
-    """
-    model = pyo.ConcreteModel(name="chemakzo")
-
-    # Set problem parameter values
-    model.k = pyo.Param([1,2,3,4], initialize={
-        1:18.7,
-        2:0.58,
-        3:0.09,
-        4:0.42})
-    model.Ke = pyo.Param(initialize=34.4)
-    model.klA = pyo.Param(initialize=3.3)
-    model.Ks = pyo.Param(initialize=115.83)
-    model.pCO2 = pyo.Param(initialize=0.9)
-    # The following parameter H, is best made a parameter, but will use a
-    # variable and constraint instead to test non-time-indexed constraints
-    #model.H = pyo.Param(initialize=737)
-
-    # Problem variables ydot = dy/dt,
-    #    (dy6/dt is not explicitly in the equations, so only 5 ydots)
-    model.H = pyo.Var(initialize=737)
-    model.t = pyodae.ContinuousSet(bounds=(0,180))
-    model.y = pyo.Var(model.t, [1,2,3,4,5], initialize=1.0)  #
-    model.y6 = pyo.Var(model.t, initialize=1.0)  #
-    model.ydot = pyodae.DerivativeVar(model.y, wrt=model.t) # dy/dt
-    model.r = pyo.Var(model.t, [1,2,3,4,5], initialize=1.0)
-    model.Fin = pyo.Var(model.t, initialize=1.0)
-
-    # Non-time indexed constraint (just for testing)
-    model.H_eqn = pyo.Constraint(expr=model.H==737)
-
-    # Equations
-    @model.Constraint(model.t)
-    def eq_ydot1(b, t):
-        if t==0:
-            pyo.Constraint.Skip
-        return b.ydot[t, 1] == -2.0*b.r[t, 1] + b.r[t, 2] - b.r[t, 3] - b.r[t, 4]
-    @model.Constraint(model.t)
-    def eq_ydot2(b, t):
-        if t==0:
-            pyo.Constraint.Skip
-        return b.ydot[t, 2] == -0.5*b.r[t, 1] - b.r[t, 4] - 0.5*b.r[t, 5] + b.Fin[t]
-    @model.Constraint(model.t)
-    def eq_ydot3(b, t):
-        if t==0:
-            pyo.Constraint.Skip
-        return b.ydot[t, 3] == b.r[t, 1] - b.r[t, 2] + b.r[t, 3]
-    @model.Constraint(model.t)
-    def eq_ydot4(b, t):
-        if t==0:
-            pyo.Constraint.Skip
-        return b.ydot[t, 4] == -b.r[t, 2] + b.r[t, 3] - 2.0*b.r[t, 4]
-    @model.Constraint(model.t)
-    def eq_ydot5(b, t):
-        if t==0:
-            pyo.Constraint.Skip
-        return b.ydot[t, 5] == b.r[t, 2] - b.r[t, 3] + b.r[t, 5]
-    @model.Constraint(model.t)
-    def eq_y6(b, t):
-        return 0 == b.Ks*b.y[t, 1]*b.y[t, 4] - b.y6[t]
-
-    @model.Constraint(model.t)
-    def eq_r1(b, t):
-        return b.r[t, 1] == b.k[1]*b.y[t, 1]**4*b.y[t, 2]**0.5
-    @model.Constraint(model.t)
-    def eq_r2(b, t):
-        return b.r[t, 2] == b.k[2]*b.y[t, 3]*b.y[t, 4]
-    @model.Constraint(model.t)
-    def eq_r3(b, t):
-        return b.r[t, 3] == b.k[2]/b.Ke*b.y[t, 1]*b.y[t, 5]
-    @model.Constraint(model.t)
-    def eq_r4(b, t):
-        return b.r[t, 4] == b.k[3]*b.y[t, 1]*b.y[t, 4]**2
-    @model.Constraint(model.t)
-    def eq_r5(b, t):
-        return b.r[t, 5] == b.k[4]*b.y6[t]**2*b.y[t, 2]**0.5
-    @model.Constraint(model.t)
-    def eq_Fin(b, t):
-        return b.Fin[t] == b.klA*(b.pCO2/b.H - b.y[t, 2])
-
-    # Set initial condtions and solve initial from the values of differntial
-    # variables (r and y6 well and the derivative vars too).
-    y0 = {1:0.444, 2:0.00123, 3:0.0, 4:0.007, 5:0.0} #initial differntial vars
-    for i in y0:
-        model.y[0, i].fix(y0[i])
-
-    discretizer = pyo.TransformationFactory('dae.finite_difference')
-    discretizer.apply_to(model, nfe=1, scheme='BACKWARD')
-
-    return (
-        model,
-        0.1150794920661702,
-        0.1203831471567715e-2,
-        0.1611562887407974,
-        0.3656156421249283e-3,
-        0.1708010885264404e-1,
-        0.4873531310307455e-2,
-    )
-
 @lru_cache(maxsize=10)
 def ipopt_has_linear_solver(linear_solver):
     """Check if IPOPT can use the specified linear solver.
